adding IObjectParser and derived classes
