+++ src/Bicep.Core/Semantics/JsonObjectParser.cs
@@ -1,0 +1,11 @@
+using SharpYaml.Serialization;
+
+namespace Bicep.Core.Semantics
+{
+    public static class JsonObjectParser : ObjectParser
+    {
+        public static JToken ExtractTokenFromObject(string fileContent)
+        {
+            return fileContent.TryFromJson<JToken>();
+        }
+}
