+++ src/Bicep.Core/Semantics/YamlObjectParser.cs
@@ -1,0 +1,12 @@
+using SharpYaml.Serialization;
+
+namespace Bicep.Core.Semantics
+{
+    public static class YamlObjectParser : ObjectParser
+    {
+        public static JToken ExtractTokenFromObject(string fileContent)
+        {
+            return JToken.FromObject(new Serializer().Deserialize(fileContent)!);
+        }
+    }
+}
